Remove redundant pre-install ZIP validation
